requirements/base.txt
django>=1.8.0
djangorestframework==3.3.2
python-telegram-bot==3.2.0